submit_form/requirements.txt
click==8.0.4
dataclasses==0.8
dominate==2.7.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.29
protobuf==3.19.4
typing-extensions==4.1.1
visitor==0.1.3
Werkzeug==2.0.3
zipp==3.6.0
